Improvements to Format layer implementation
